chore: move contents of `witness/btc_e.rs` to `btc.rs`.
